refactor(broker): extract inner class
