+++ broker/src/main/java/io/camunda/zeebe/broker/system/partitions/impl/NewPartitionTransitionImpl.java
@@ -19,7 +19,6 @@ import io.camunda.zeebe.util.sched.ConcurrencyControl;
 import io.camunda.zeebe.util.sched.future.ActorFuture;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Stack;
 import org.slf4j.Logger;
 
 public final class NewPartitionTransitionImpl implements PartitionTransition {
@@ -29,9 +28,9 @@ public final class NewPartitionTransitionImpl implements PartitionTransition {
   private final List<PartitionTransitionStep> steps;
   private PartitionTransitionContext context;
   private ConcurrencyControl concurrencyControl;
-  private Transition lastTransition;
+  private PartitionTransitionProcess lastTransition;
   // these two should be set/cleared in tandem
-  private Transition currentTransition;
+  private PartitionTransitionProcess currentTransition;
   private ActorFuture<Void> currentTransitionFuture;
   // these two should be set in tandem
 
@@ -119,7 +118,8 @@ public final class NewPartitionTransitionImpl implements PartitionTransition {
 
   private void startNewTransition(
       final ActorFuture<Void> nextTransitionFuture, final long term, final Role role) {
-    currentTransition = new Transition(steps, concurrencyControl, context, term, role);
+    currentTransition =
+        new PartitionTransitionProcess(steps, concurrencyControl, context, term, role);
     currentTransitionFuture = nextTransitionFuture;
     concurrencyControl.runOnCompletion(
         currentTransitionFuture,
@@ -129,142 +129,5 @@ public final class NewPartitionTransitionImpl implements PartitionTransition {
           currentTransitionFuture = null;
         });
     currentTransition.start(currentTransitionFuture);
-  }
-
-  private static final class Transition {
-
-    private final List<PartitionTransitionStep> pendingSteps;
-    private final Stack<PartitionTransitionStep> startedSteps = new Stack<>();
-    private final ConcurrencyControl concurrencyControl;
-    private final PartitionTransitionContext context;
-    private final long term;
-    private final Role role;
-    private boolean cancelRequested = false;
-
-    private Transition(
-        final List<PartitionTransitionStep> pendingSteps,
-        final ConcurrencyControl concurrencyControl,
-        final PartitionTransitionContext context,
-        final long term,
-        final Role role) {
-      this.pendingSteps = new ArrayList<>(pendingSteps);
-      this.concurrencyControl = concurrencyControl;
-      this.context = context;
-      this.term = term;
-      this.role = role;
-    }
-
-    private void start(final ActorFuture<Void> future) {
-      LOG.info(format("Transition to %s on term %d starting", role, term));
-
-      if (pendingSteps.isEmpty()) {
-        LOG.info("No steps defined for transition");
-        future.complete(null);
-        return;
-      }
-
-      proceedWithTransition(future);
-    }
-
-    private void proceedWithTransition(final ActorFuture<Void> future) {
-      if (cancelRequested) {
-        LOG.info(format("Cancelling transition to %s on term %d", role, term));
-        future.complete(null);
-        return;
-      }
-
-      concurrencyControl.submit(
-          () -> {
-            final var nextStep = pendingSteps.remove(0);
-            startedSteps.push(nextStep);
-
-            LOG.info(
-                format(
-                    "Transition to %s on term %d - executing %s", role, term, nextStep.getName()));
-
-            nextStep
-                .transitionTo(context, term, role)
-                .onComplete((nil, error) -> onStepCompletion(future, error));
-          });
-    }
-
-    private void onStepCompletion(final ActorFuture<Void> future, final Throwable error) {
-      if (error != null) {
-        LOG.error(error.getMessage(), error);
-        future.completeExceptionally(error);
-
-        return;
-      }
-
-      if (pendingSteps.isEmpty()) {
-        LOG.info(format("Transition to %s on term %d completed", role, term));
-        future.complete(null);
-
-        return;
-      }
-
-      proceedWithTransition(future);
-    }
-
-    public ActorFuture<Void> cleanup(final long newTerm, final Role newRole) {
-      LOG.info(
-          format(
-              "Cleanup of transition to %s on term %d starting (in preparation for new transition to %s)",
-              role, term, newRole));
-      final ActorFuture<Void> cleanupFuture = concurrencyControl.createFuture();
-
-      if (startedSteps.isEmpty()) {
-        LOG.info("No steps to clean up");
-        cleanupFuture.complete(null);
-      } else {
-        proceedWithCleanup(cleanupFuture, newTerm, newRole);
-      }
-      return cleanupFuture;
-    }
-
-    private void proceedWithCleanup(
-        final ActorFuture<Void> future, final long newTerm, final Role newRole) {
-      concurrencyControl.submit(
-          () -> {
-            final var nextCleanupStep = startedSteps.pop();
-
-            LOG.info(
-                format(
-                    "Cleanup of transition to %s on term %d - executing %s",
-                    role, term, nextCleanupStep.getName()));
-
-            nextCleanupStep
-                .prepareTransition(context, newTerm, newRole)
-                .onComplete(
-                    (nil, error) -> onCleanupStepCompletion(future, error, newTerm, newRole));
-          });
-    }
-
-    private void onCleanupStepCompletion(
-        final ActorFuture<Void> future,
-        final Throwable error,
-        final long newTerm,
-        final Role newRole) {
-      if (error != null) {
-        LOG.error(error.getMessage(), error);
-        future.completeExceptionally(error);
-
-        return;
-      }
-
-      if (startedSteps.isEmpty()) {
-        LOG.info(format("Cleanup of transition to %s on term %d completed", role, term));
-        future.complete(null);
-
-        return;
-      }
-
-      proceedWithCleanup(future, newTerm, newRole);
-    }
-
-    private void cancel() {
-      LOG.info(format("Received cancel signal for transition to %s on term %d", role, term));
-      cancelRequested = true;
-    }
   }
 }
